Refactor entity behaviour to component behaviour + make component trigger-able by any generators
This allows to apply the component system_command to any entity instance. The component is an action and can be triggered (input property: trigger) externally or internally. Entity instances which are also a generator (output property: trigger) automatically triggers internally and are executing the system command.

For example a ScheduledJob is a generator and can be extended with the component system_command. The scheduled job triggers via the property trigger. The component system_command is an action and gets triggered by the same property and executes the command each time the scheduled job fires.